Resolve md678685s comments
Thx!
