+++ Essentials/src/com/earth2me/essentials/signs/SignSell.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 import static com.earth2me.essentials.I18n.tl;
 
 
-public class SignSell extends EssentialsSign {
+public class SignSell extends EssentialsShopSign {
     public SignSell() {
         super("Sell");
     }
@@ -21,10 +21,9 @@ public class SignSell extends EssentialsSign {
     @Override
     protected boolean onSignCreate(final ISign sign, final User player, final String username, final IEssentials ess) throws SignException {
         validateTrade(sign, 1, 2, player, ess);  
-        if(sign.getLine(3).isEmpty()) {
-            updateFromWorth(sign, player, ess, new BigDecimal("1.0"));
-        }
-        else {
+        if (sign.getLine(3).isEmpty()) {
+            updateFromWorth(sign, player, ess);
+        } else {
             validateTrade(sign, 3, ess);
         }
         return true;
@@ -32,7 +31,7 @@ public class SignSell extends EssentialsSign {
 
     @Override
     protected boolean onSignInteract(final ISign sign, final User player, final String username, final IEssentials ess) throws SignException, ChargeException, MaxMoneyException {
-        if(sign.getLine(3).startsWith(tl("signFormatWorth")) && updateFromWorth(sign, player, ess, new BigDecimal("1.0"))) {
+        if (sign.getLine(3).startsWith(tl("signFormatWorth")) && updateFromWorth(sign, player, ess)) {
             return false;
         }
         Trade charge = getTrade(sign, 1, 2, player, ess);
