+++ Essentials/src/messages.properties
@@ -597,7 +597,6 @@ powerToolRemove=\u00a76Command \u00a7c{0}\u00a76 removed from \u00a7c{1}\u00a76.
 powerToolRemoveAll=\u00a76All commands removed from \u00a7c{0}\u00a76.
 powerToolsDisabled=\u00a76All of your power tools have been disabled.
 powerToolsEnabled=\u00a76All of your power tools have been enabled.
-priceChanged=\u00a76The price of \u00a7c{0}x {1}\u00a76 has changed from \u00a7c{2}\u00a76 to \u00a7c{3}\u00a76.
 powertoolCommandDescription=Assigns a command to the item in hand.
 powertoolCommandUsage=/<command> [l:|a:|r:|c:|d:][command] [arguments] - {player} can be replaced by name of a clicked player.
 powertooltoggleCommandDescription=Enables or disables all current powertools.
@@ -606,7 +605,7 @@ ptimeCommandDescription=Adjust player's client time. Add @ prefix to fix.
 ptimeCommandUsage=/<command> [list|reset|day|night|dawn|17:30|4pm|4000ticks] [player|*]
 pweatherCommandDescription=Adjust a player's weather
 pweatherCommandUsage=/<command> [list|reset|storm|sun|clear] [player|*]
-priceChanged=\u00a76The price of \u00a7c{0}x {1}\u00a76 has changed from \u00a7e{2}\u00a76 to \u00a7e{3}\u00a76.
+priceChanged=\u00a76The price of \u00a7c{0}x {1}\u00a76 has changed from \u00a7e{2}\u00a76 to \u00a7e{3}\u00a76. Please click the sign again to retry the transaction.
 pTimeCurrent=\u00a7c{0}\u00a76''s time is\u00a7c {1}\u00a76.
 pTimeCurrentFixed=\u00a7c{0}\u00a76''s time is fixed to\u00a7c {1}\u00a76.
 pTimeNormal=\u00a7c{0}\u00a76''s time is normal and matches the server.
